Update default React app boilerplate to OASIS Web UI
The default React App has been replaced with OASIS Web app layout. This includes transition from a simple counter application to a layout with input fields and display areas. Bootstrap CSS and components were added to improve UI, and the previous static CSS was removed. This is an iteration towards creating a complete web app for Algebraic solutions.